refactor(auth): policy 네이밍을 helper로 변경
