+++ src/auth/application/helpers/auth-result.helper.ts
