+++ src/auth/application/usecases/switch-user.usecase.ts
@@ -6,7 +6,7 @@ import type { AuthSessionPort } from '../ports/auth-session.port';
 import { AuthError } from '../errors/auth.error';
 import { AuthSessionOutput } from '../dtos/auth-session-output.dto';
 import { AuthPlatform } from 'src/shared/types/auth-platform.type';
-import { issueAuthResult } from '../policies/auth-result.policy';
+import { issueAuthResult } from '../helpers/auth-result.helper';
 
 @Injectable()
 export class SwitchUserUseCase {
