+++ src/auth/presentation/strategies/google.strategy.ts
@@ -9,7 +9,7 @@ import { OAuthUser } from '../../domain/auth.types';
 import {
   buildOAuthState,
   parseOAuthState,
-} from '../policies/oauth-state.policy';
+} from '../helpers/oauth-state.helper';
 
 @Injectable()
 export class GoogleStrategy extends PassportStrategy(Strategy, 'google') {
