+++ src/clips/application/helpers/clip-data.helper.ts
