+++ src/clips/application/helpers/clip-type-detector.spec.ts
