+++ src/clips/application/helpers/clip-type-detector.ts
