+++ src/clips/application/usecases/create-clip.usecase.ts
@@ -4,7 +4,7 @@ import type { Clip } from '../../domain/clip.types';
 import type { ClipsRepository } from '../../domain/clips.repository';
 import { CreateClipInput } from '../dtos/create-clip-input.dto';
 import { ClipsError } from '../errors/clips.error';
-import { resolveClipData } from '../policies/clip-data.policy';
+import { resolveClipData } from '../helpers/clip-data.helper';
 
 @Injectable()
 export class CreateClipUseCase {
