+++ src/clips/application/usecases/list-clips.helper.ts
