+++ src/clips/application/usecases/list-favorite-clips.usecase.ts
@@ -14,7 +14,7 @@ import {
 import {
   resolveClipSearchTarget,
   validateClipCursor,
-} from './list-clips.policy';
+} from './list-clips.helper';
 
 @Injectable()
 export class ListFavoriteClipsUseCase {
