+++ src/clips/application/usecases/list-recent-clips.usecase.ts
@@ -12,7 +12,7 @@ import { normalizeCursor, normalizeType } from './list-clips.common';
 import {
   listRecentClipsWithLikedPriority,
   resolveRecentClipSearchTarget,
-} from './list-clips.policy';
+} from './list-clips.helper';
 
 @Injectable()
 export class ListRecentClipsUseCase {
