+++ src/users/application/usecases/update-me.usecase.ts
@@ -3,12 +3,12 @@ import { USERS_REPOSITORY } from '../../domain/users.repository';
 import type { UsersRepository } from '../../domain/users.repository';
 import { UpdateMeInput } from '../dtos/update-me-input.dto';
 import { UserProfileOutput } from '../dtos/user-profile-output.dto';
-import { mapMeResponse } from '../policies/map-me-response.policy';
+import { mapMeResponse } from '../helpers/map-me-response.helper';
 import { UsersError } from '../errors/users.error';
 import {
   buildUpdateMeParams,
   resolveCurrentAuthAccount,
-} from '../policies/update-me.policy';
+} from '../helpers/update-me.helper';
 
 @Injectable()
 export class UpdateMeUseCase {
