+++ src/workspaces/application/helpers/subscription-response.helper.ts
