+++ src/workspaces/application/usecases/get-my-subscription.usecase.ts
@@ -2,8 +2,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import { WORKSPACES_REPOSITORY } from '../../domain/workspaces.repository';
 import type { WorkspacesRepository } from '../../domain/workspaces.repository';
 import { MySubscriptionOutput } from '../dtos/my-subscription-output.dto';
-import { normalizeExpiredSubscription } from '../policies/subscription-expiration.policy';
-import { toMySubscriptionResponse } from '../policies/subscription-response.policy';
+import { normalizeExpiredSubscription } from '../helpers/subscription-expiration.helper';
+import { toMySubscriptionResponse } from '../helpers/subscription-response.helper';
 
 @Injectable()
 export class GetMySubscriptionUseCase {
